Refactor Playwright test
