+++ tmail_selenium/pom.xml
@@ -23,8 +23,7 @@
 		<dependency>
 			<groupId>junit</groupId>
 			<artifactId>junit</artifactId>
-			<version>4.11</version>
-			<scope>test</scope>
+			<version>4.13.2</version>
 		</dependency>
 
 		<dependency>
