+++ tmail_selenium/src/main/java/com/tmail/base/CoreRobot.java
@@ -1,4 +1,4 @@
-package com.linagora.tmail.base;
+package com.tmail.base;
 
 import com.microsoft.playwright.Page;
