+++ tmail_selenium/src/main/java/com/tmail/base/SupportedPlatform.java
@@ -1,4 +1,4 @@
-package com.linagora.tmail.base;
+package com.tmail.base;
 
 public enum SupportedPlatform {
     CHROME,
