+++ tmail_selenium/src/main/java/com/tmail/robots/MailboxDashboardRobot.java
@@ -1,11 +1,11 @@
-package com.linagora.tmail.robots;
+package com.tmail.robots;
 
 import static org.junit.Assert.assertTrue;
 
 import java.util.regex.Pattern;
 
-import com.linagora.tmail.base.CoreRobot;
 import com.microsoft.playwright.Page;
+import com.tmail.base.CoreRobot;
 
 public class MailboxDashboardRobot extends CoreRobot {
