+++ tmail_selenium/src/main/java/com/tmail/scenarios/ComposeEmailScenario.java
@@ -1,20 +1,20 @@
-package com.linagora.tmail.usecases;
+package com.tmail.scenarios;
 
 import java.util.ArrayList;
 import java.util.List;
 
-import com.linagora.tmail.base.UseCase;
-import com.linagora.tmail.robots.ComposerRobot;
-import com.linagora.tmail.robots.MailboxDashboardRobot;
 import com.microsoft.playwright.Page;
+import com.tmail.base.BaseScenario;
+import com.tmail.robots.ComposerRobot;
+import com.tmail.robots.MailboxDashboardRobot;
 
-public class ComposeEmailUseCase extends UseCase {
+public class ComposeEmailScenario extends BaseScenario {
     String testUrl;
     String username;
     String password;
     String additionalReceipent;
 
-    public ComposeEmailUseCase(String testUrl, String username, String password, String additionalReceipent) {
+    public ComposeEmailScenario(String testUrl, String username, String password, String additionalReceipent) {
         this.testUrl = testUrl;
         this.username = username;
         this.password = password;
@@ -26,7 +26,7 @@ public class ComposeEmailUseCase extends UseCase {
         MailboxDashboardRobot mailboxDashboardRobot = new MailboxDashboardRobot(page);
         ComposerRobot composerRobot = new ComposerRobot(page);
 
-        LoginUseCase loginUseCase = new LoginUseCase(testUrl, username, password);
+        LoginScenario loginUseCase = new LoginScenario(testUrl, username, password);
         loginUseCase.execute(page);
 
         mailboxDashboardRobot.openComposer();
